pyproject.toml
[project]
name = "S15A"
version = "0.2.0"
description = "Agentic AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.3",
    "asyncio>=3.4.3",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.10.0",
    "httpx>=0.28.1",
    "llama-index>=0.12.31",
    "llama-index-embeddings-google-genai>=0.1.0",
    "markitdown[all]>=0.1.1",
    "mcp[cli]>=1.6.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "pymupdf4llm>=0.0.21",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
    "jinja2>=3.1.6",
    "playwright>=1.52.0",
    "readability-lxml>=0.8.4.1",
    "networkx>=3.4.2",
    "json-repair>=0.44.1",
    "markdownify>=1.1.0",
    "uvicorn>=0.34.2",
    "fastapi>=0.115.12",
    "sse-starlette>=2.3.4",
    "feedparser>=6.0.12",
    "beautifulsoup4>=4.13.4",
    "aiohttp>=3.11.18",
    "watchdog>=3.0.0",
    "langchain-ollama>=0.3.2",
    "langchain-google-genai>=2.1.2",
    "rank_bm25>=0.2.2",
    "apscheduler>=3.11.2",
    "psutil>=7.2.2",
    "qdrant-client>=1.7.0",
    "fastembed>=0.4.0",
    "python-pptx>=1.0.2",
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "python-multipart>=0.0.9",
    "markdown>=3.6",
    "xhtml2pdf>=0.2.16",
    "neo4j>=5.14.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "pymongo>=4.10.0",
    "ray[default]==2.40.0",
    "memo>=0.2.4",
    "mem0ai>=1.0.4",
    "sounddevice>=0.5.5",
    "pvporcupine==4.0.2",
    "openwakeword>=0.6.0",
    "pytest-asyncio>=1.3.0",
    "tensorflow>=2.20.0",
    "pytest>=9.0.2",
    "faster-whisper>=1.2.1",
    "numpy>=2.4.0",
    "noisereduce>=3.0.3",
    "websocket-client>=1.9.0",
    "azure-cognitiveservices-speech>=1.48.2",
    "piper-tts>=1.2.0",
    "pynacl>=1.6.2",
    "mongomock>=4.3.0",
    "sqlmodel",
    "alembic",
    "pyjwt",
    "passlib",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "chardet<6",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Explicitly list Python packages to avoid accidental inclusion of non-package directories
packages = [
    "agents",
    "config",
    "core",
    "mcp_servers",
    "memory",
    "ops",
    "remme",
    "routers",
    "shared",
    "tools",
    "ui",
    "canvas",
    "gateway_api",
    "gateway_api.connectors",
    "storage",
    "session",
]

[tool.uv]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0",
]

# ---------------------------------------------------------------------------
# Ruff — fast Python linter (replaces flake8/isort/pyupgrade)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 120
target-version = "py311"
include = [
    "agents/**/*.py",
    "channels/**/*.py",
    "core/**/*.py",
    "config/**/*.py",
    "gateway/**/*.py",
    "shared/**/*.py",
    "tests/**/*.py",
]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "platform-frontend",
    "data",
]

[tool.ruff.lint]
# E/W = pycodestyle, F = pyflakes, I = isort, UP = pyupgrade
select = ["E", "W", "F", "I", "UP"]
ignore = [
    "E501",   # line too long — handled by line-length above
    "F401",   # imported but unused — Ray remote classes import side-effects
    "E402",   # module level import not at top — Ray actors use deferred imports
]

[tool.ruff.lint.per-file-ignores]
# base_agent.py is legacy code — suppress cosmetic whitespace and style errors
"agents/base_agent.py" = ["W291", "W293", "I001", "UP007", "UP042", "UP045"]

# ---------------------------------------------------------------------------
# Mypy — static type checker
# Only checks agents/ (P08 scope) to avoid false-positives in legacy code
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
files = ["agents"]
explicit_package_bases = true   # fixes "found twice" when run from project root
follow_imports = "silent"        # silence errors in imported legacy files (remme/, core/, etc.)
ignore_missing_imports = true    # Ray stubs and internal packages are untyped
warn_unused_ignores = false
check_untyped_defs = true        # Check bodies even without full annotations
# Relax strictness for pragmatic compatibility with existing codebase
disallow_untyped_defs = false
disallow_incomplete_defs = false
[[tool.mypy.overrides]]
module = ["ray.*", "google.*", "mcp.*", "PIL.*", "psutil", "yaml"]
ignore_missing_imports = true